image/requirements.txt
mutagen==1.45.1
requests==2.24.0
image_slicer==2.1.1
# Pillow==8.1.2
# Pillow==7.2.0
Pillow>=8.2.0
